% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gcd}
\alias{gcd}
\title{The greatest common divisor of two numbers.}
\usage{
gcd(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The greatest common divisor of \code{x} and \code{y}.
}
\description{
The greatest common divisor of two numbers.
}
\examples{
gcd(12, 16)
gcd(3, 5)
}

