% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{lcm}
\alias{lcm}
\title{The least common multiple of two numbers.}
\usage{
lcm(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The least common multiple of \code{x} and \code{y}.
}
\description{
The least common multiple of two numbers.
}
\examples{
lcm(12, 16)
lcm(3, 5)
}

