% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{pop.sd}
\alias{pop.sd}
\title{Population standard deviation.}
\usage{
pop.sd(d)
}
\arguments{
\item{d}{A vector}
}
\value{
\code{sd()} without the \code{n - 1} division
}
\description{
Population standard deviation.
}
\examples{
pop.sd(1:10)
}

