% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{pop.var}
\alias{pop.var}
\title{Population variance.}
\usage{
pop.var(d)
}
\arguments{
\item{d}{A vector}
}
\value{
\code{var()} without the \code{n - 1} division
}
\description{
Population variance.
}
\examples{
pop.var(1:10)
}

